reorg change-atom-id
