+++ packages/dds/tree/src/core/rebase/types.ts
@@ -5,7 +5,7 @@
 
 import { assert } from "@fluidframework/core-utils";
 import { isStableId, StableId } from "@fluidframework/id-compressor";
-import { Brand, NestedMap, RangeMap, brandedStringType, generateStableId } from "../../util";
+import { Brand, brandedStringType, generateStableId } from "../../util";
 
 /**
  * The identifier for a particular session/user/client that can generate `GraphCommit`s
@@ -22,55 +22,6 @@ export const SessionIdSchema = brandedStringType<SessionId>();
 export type RevisionTag = StableId;
 export type EncodedRevisionTag = Brand<string, "EncodedRevisionTag">;
 export const RevisionTagSchema = brandedStringType<EncodedRevisionTag>();
-
-/**
- * An ID which is unique within a revision of a `ModularChangeset`.
- * A `ModularChangeset` which is a composition of multiple revisions may contain duplicate `ChangesetLocalId`s,
- * but they are unique when qualified by the revision of the change they are used in.
- * @internal
- */
-export type ChangesetLocalId = Brand<number, "ChangesetLocalId">;
-
-/**
- * A globally unique ID for an atom of change, or a node associated with the atom of change.
- * @internal
- *
- * @privateRemarks
- * TODO: Rename this to be more general.
- */
-export interface ChangeAtomId {
-	/**
-	 * Uniquely identifies the changeset within which the change was made.
-	 * Only undefined when referring to an anonymous changesets.
-	 */
-	readonly revision?: RevisionTag;
-	/**
-	 * Uniquely identifies, in the scope of the changeset, the change made to the field.
-	 */
-	readonly localId: ChangesetLocalId;
-}
-
-export interface EncodedChangeAtomId {
-	readonly revision?: EncodedRevisionTag;
-	readonly localId: ChangesetLocalId;
-}
-
-/**
- * @internal
- */
-export type ChangeAtomIdMap<T> = NestedMap<RevisionTag | undefined, ChangesetLocalId, T>;
-
-/**
- * @internal
- */
-export type ChangeAtomIdRangeMap<T> = Map<RevisionTag | undefined, RangeMap<T>>;
-
-/**
- * @returns true iff `a` and `b` are the same.
- */
-export function areEqualChangeAtomIds(a: ChangeAtomId, b: ChangeAtomId): boolean {
-	return a.localId === b.localId && a.revision === b.revision;
-}
 
 /**
  * @returns a `RevisionTag` from the given string, or fails if the string is not a valid `RevisionTag`
