+++ packages/dds/tree/src/feature-libraries/change-atom-id/index.ts
@@ -1,0 +1,14 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export {
+	ChangesetLocalId,
+	ChangeAtomId,
+	ChangeAtomIdRangeMap,
+	ChangeAtomIdMap,
+} from "./changeAtomIdTypes";
+export { ChangesetLocalIdSchema, EncodedChangeAtomId } from "./changeAtomIdFormat";
+export { encodeChangeAtomId, decodeChangeAtomId } from "./changeAtomIdCodec";
+export { areEqualChangeAtomIds } from "./changeAtomIdUtils";
