+++ packages/dds/tree/src/feature-libraries/optional-field/optionalFieldChangeTypes.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { ChangeAtomId } from "../../core";
+import { ChangeAtomId } from "../change-atom-id";
 import { NodeChangeset } from "../modular-schema";
 
 /**
